refactor: support 패키지 대신 application 패키지 하위로 이동
